Move login/logout methods to their own test class
Login (BeforeMethod) and Login (AfterMethod) methods are repeated in
all tests that are not LoginTests. BaseTests_LoginLogout inherits from BaseTests
and the non-LoginTests will inherit from BaseTests_LoginLogout. Whereas LoginTests
will continue to inherit from BaseTests since we do not want login and logout methods
to be performed before and after these tests.
